Restructure profiles for clarity.
