+++ src/main/java/dev/cheem/projects/invoicestore/config/profile/prod/DatabaseConfigProd.java
@@ -1,6 +1,5 @@
-package dev.cheem.projects.invoicestore.config;
+package dev.cheem.projects.invoicestore.config.profile.prod;
 
-import dev.cheem.projects.invoicestore.model.DatabaseInstance;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
