+++ src/main/resources/application-debug.properties
@@ -1,3 +1,2 @@
 # Logging
 logging.level.dev.cheem.projects.invoicestore=debug
-logging.level.io.findify.s3mock=debug
