+++ src/main/resources/application-prod.properties
@@ -1,6 +1,3 @@
-# Logging
-logging.level.root=warn
-logging.level.dev.cheem.projects.invoicestore=warn
 # DB
 spring.datasource.platform=postgres
 spring.datasource.url=jdbc:postgresql://${JPA_URL}/${JPA_DB}?binaryTransfer=false
