+++ src/main/resources/application-warn-all.properties
@@ -1,0 +1,4 @@
+# Logging
+logging.level.root=warn
+logging.level.org.springframework=warn
+logging.level.dev.cheem.projects.invoicestore=warn
