+++ src/main/resources/application.properties
@@ -1,5 +1,8 @@
 # Logging
 logging.level.io.netty.util=ERROR
+logging.level.root=info
+logging.level.org.springframework=info
+logging.level.dev.cheem.projects.invoicestore=info
 # JPA
 spring.jpa.open-in-view=false
 # Static
@@ -25,9 +28,6 @@ spring.servlet.multipart.max-request-size=215MB
 # Endpoint for GraphQL Server, see graphql-java-spring-boot-starter-webmvc for more details.
 graphql.url=api/graphql
 
-# Logging
-logging.level.root=info
-logging.level.dev.cheem.projects.invoicestore=info
 # DB
 spring.datasource.driverClassName=org.h2.Driver
 # JPA
